chore(driver): make force_tp_set as robust as possible.
It now **can't** have desynced `g_tracepoints_attached` and `g_tracepoints_refs`.
Only thing that can happens is that when destroying a consumer (that is the only one present),
for any reason, not all tracepoints can be detached; so we end up with some refs/attached tracepoints
that are not properly cleaned, even if no consumer exists.

This just means that we won't be able to disable them anymore because we will have
some zombie references to those tracepoints.
Not a big deal, and i think it should really never happen.

Signed-off-by: Federico Di Pierro <nierro92@gmail.com>